background on and key handling
the background has been set to the default tile, will need to add a background string variable and an arry of arry's that will let us have the back ground change based on that. also need to have collision added.
finaly I also added all of the key inputs so all that is left to do is fill out the respective methods.